:recycle: Moves everything in a lib folder for better readability
