+++ lib/aabb.rb
