+++ lib/color.rb
