+++ lib/engines/core.rb
