+++ lib/fsm.rb
