+++ lib/input.rb
