+++ lib/interfaces/uses_mouse.rb
