+++ lib/node.rb
