+++ lib/nodes/kinematic_body.rb
