+++ lib/nodes/label.rb
