+++ lib/nodes/node_2d.rb
