+++ lib/nodes/rectangle_shape.rb
