+++ lib/nodes/shape.rb
