+++ lib/nodes/sprite.rb
