+++ lib/nodes/static_body.rb
