+++ lib/root.rb
