+++ lib/scene.rb
