+++ lib/vector2.rb
